Scale networks to 4B synapses
Use unsigned instead of int for the synapse index, allowing us to
instantiate networks with up to ~4.3B synapses (given enough memory).

+++ spice/models/brunel.h
@@ -37,7 +37,7 @@ struct brunel : model
 			int const Tref = 20;              // dt
 			float const Vthres = 0.02f;       // v
 
-			if( n.id() < info.num_neurons / 2 ) // poisson neuron
+			if( n.id() < static_cast<unsigned>( info.num_neurons / 2 ) ) // poisson neuron
 			{
 				float const firing_rate = 20; // Hz
